requirements.txt
# Core framework
google-adk==1.2.1
google-genai==1.18.0
mcp==1.9.2
litellm==1.72.1
python-dotenv==1.1.0

# Stdio servers (ChromaDB + embeddings)
langchain-community
langchain-ollama
langchain-chroma
langchain-experimental
langchain-openai

# Optional integrations used by tools/examples
langchain-zotero-retriever
pyzotero
beautifulsoup4
mcp-server-fetch
